allow autocompleting username when adding user to organisation
It seems this was explicitely disabled by setting access to `user_list` as False. However, autocompleting the user name is desired. As it is unclear why `user_list` was restricted in access, this commit just changes the access for `user_autocomplete` and leaves `user_list` authorization as is.

However, this disabling of `user_list` access might be code that can just be deleted.

(by default access to `user_autocomplete` should work. It is this `user_list` custom function that specifically breaks it).