cleanup all helpers related to language and translation
introduce a new helper `lang_text` which should be the default to use for any dicts that have as key a language.
The alternatives used in the app now either do no fallback at all or do the one of scheming plugin, which only falls back to english.

some helpers are renamed to have the `benap` prefix to make clear they are from this plugin and not another one.