improve translations flow
Before you had to specify with a comment which languages you wanted to overwrite the ckan core translation. Then, when editing the .po file, you'd need to take in mind the comment above for this.

Now, the .po file will only include the translations from the overwrite file where the language is specified. This makes the process much easier.

This is a step to help make this more user-friendly for external translation help.