Don't truncate html, resulting in potential missing closing tags (like </i>)
Ckan-core removes the markdown when rendering here, because it truncates the text.
benap-plugin instead renders the markdown and then truncates. This means that the code to end italic might be truncated away, creating the mentioned bug.

Alternatively is first truncating and then rendering the markdown. However, if truncation is done e.g. *my text I want in italic and...., it will not show in italic, but instead show that exact text (because the ending * is truncated away).

This is not the desired result, but better than the error of a missing html tag.