translate facet titles via IFacets interface
Before there was a confusing mix between a brittle translation on the template, some helper that translated some of the functions, but also a usage of IFacets.

Instead, everything is now done via the IFacets interface, as the ckan docs also recommend. There is no need for custom code for this, as it is build in.