fix file deletion when uploading new file
files were not deleted when a new file was uploaded. This meant the server might keep files a user incorrectly uploaded and did not want to be kept.
Delete these files correctly now with a hack, as the upload code is far too old to easily fix.

Note that this patch is not done for logo upload: an old logo is still kept in file storage. A similar fix could be used if needed.