fix: allow not passing optional multi language fields
fluent_text can't handle missing strings or empty string. To easily fix this, default to an empty dict if the value is missing. This will be parsed correctly and has the same meaning.

Note: some other fields use some convoluted fixes for this. Those are not changed in this commit.