remove convert_validation_list_to_JSON helper
The history of this helper is dubious. There is invalid data in the database, that this helper helps to view. However, that data should have just been converted before persisting instead. This is a hacky patchjob.

Remove this helper and instead migrate the data in the database.
This means converting text to form ["...","..."]

Faulty instances are limited, so this is done by hand.