remove inline <script> for leaflet usage
- Move two script tags using leaflet to assets *and* add leaflet as assets (via files). Create webassets to load when needing map-form or map-display. This way, leaflet is only loaded when needed.
This also means able to remove the script package loading from base.html

The same leaflet version is used as was loaded. This is old and should be updated, but that is out of scope of this commit.

- BUGFIX for map-form: before, you could add multiple sections, but only last one would be saved + before if you removed a section, the input would be cleared, even if multiple other sections might be displayed on the map
Now: remove other sections when creating a new section. This way, if a section is shown on the map, it is also the section that will be saved.