style-tag cleanup: move cookies modal css to css file
this has some css classes that are not specifically for the cookies modal, so it is already a mess.
For now just move it. The css class is always loaded anyway.

In the future, this should really be split up in specific css for the cookies overlay and general css