script-tag cleanup at form_tags.html
Export to a ckan module and load the script data like that.
A choice was made to have the module be called on every input, instead of on the form input and then looping over every input.
As the assumption can be made to have the id be unique, this is okay.

This could be adjusted in case of a rewrite/feature enhancement of this function though.