script-tag cleanup for nested_field_license.html
Very coupled code, but at least it is now in a module. Note that this is coupled directly to the specifications in benap_schema (the nested_fields). So it is a bit of a mess.

Also note no functionality change with this move:
The required is set but does not add a * in the UI. This is the same as it was before. (note that there is some help_text specifying it is required when license is set to "Other")
This should also really contain documentation what it does (out of scope of this commit).