cleanup script-tag in language_selector.html
- removes unnecesary data-module-. Originally these were added as the select was using data-module="autocomplete". (https://github.com/belgium-its-steering-committee/ckanext-benap/commit/0a9b96cd88e269ee8e0d3928aa328805816f08bf) This has since changed, but it was never correctly removed.

Remove the js logic all together: all it did was sort the inputed languages. This can just be done with a simple helper .

Helper could work more generally (use .env for order and maybe handle extra locales too), but at this point this is not needed: NAP's possible language is set in stone.