cleanup script tag in datetime_tz.html
Note that here field information was set, hard-coded to a specific field_name and done with string manipulation in javascript.

As this is meant as a default value while loading the template, this can be done via the template instead.

A new field "default_to_today" is added that can be set inside the schemes. If so, it will default to the current date if no date existed. This allows removing the javascript too.

One of the controls still needs to be hidden. A small helper module is created for this.
This module simply toggles a specific element via its id.
This could be used in other places too, as there are quite a few modules that hide something (with a very specific selector).

Datetime (now) is now gotten from python, which means from the server. Because of this, the timezone is not anymore automatically set to that of the user.

For NAP's purposes, we can just always use Belgian time: Even if a user is not inside the belgium timezone, the datetimes saved to the database are expected to be in the belgium timezone, not in the local timezone of the user (as it won't convert correctly then)