Search in translations of titles when searching for organizations
Before when querying on the list of organizations, only those with the term inside the name, title or description were returned.

Adjust to also search inside the translated title field.

this search does not use solr, so the action has to be adjusted.
It does a custom fitlering first and then calls the original action with this extra filtering, to avoid having to write sorting,pagination etc logic.