use .env variable to activate google analytics
needs .env variable `CKAN__GOOGLE__TAG_ID` and `CKAN__COOKIE_DOMAIN`

Includes fixes, because previously cookies were not actually deleted correctly.
Now the domain should be passed and this is used to correctly delete cookies.
Note that for _gat_gtag_, a replace is needed from - to _, because the cookie set seems
to do this...

This isn't the perfect solution, but is a patch fix, to avoid fully rewriting this code.
Note that in practice deleting the cookies should not be strictly necessary, as setting
`ga-disable-` to true should already disable tracking from the gtags.js code.