Search in translated descriptions for organization search
Also search in description_translated when using organization_list. Note that to keep this simple, this field is seen as just a plain text.
However, the field will always include the language tags. To avoid showing everything when the search is such a lang tag, don't search in description_translated for these cases.