set date nap_checked set to true + show in pill
Save the date ("today") when nap_checked is set to true (only when it changed). Show this date on the nap_check_pill.
Note that the date to show is one lower than the year the check was performed.
So if the check was performed in 2025, the year that should be shown is 2024.

Some fiddling is done to keep this working correctly:
input of a date is always converted to "true", so the logic only has to handle true/false cases.
Only when the state was previously false and it is changed to true, will the date be edited.

This way a request setting the date will still work.
Requests trying to set a different date will just be ignored (they are seen as just passing "true", which is pretty much equivalent in terms of what the API user probably meant).

when reading the value, the date is returned, or converted to an actual boolean if it is a boolean-string.

Some edge cases have to be catched because the data might contain "true" for nap_checked for old data.

For any new data, it is either False or a date.

Even better would be to have it to be either a date or empty, but that can only be done if all "true" values are removed.

In this case, any special logic to handle nap_checked as a boolean can be removed, greatly simplifying the implemenation.

For now, it is done like this too support this boolean-style too.