show declaration fields on detail page for org member/editor
Before it checked for access to update the organization, but members and editors can't do this, even though they should see this data.
in benap/logic/action.py the organization_show action is adjusted to correctly do this:
the fields (popList) that should only be viewable by napcontrolbody/org members are popped if the user is not any of those.
This means the data is correctly returned (only what should be viewable is returned). However, it was never displayed.

Instead of checking for explicit action access, the code is changed to just check if the backend returned the field or not.
If it returned the field, it means the user had access, otherwise not.

Note that you could not do this in general: a field might not exist in the data, because it is empty, which means we still want to show it as "empty".
This only works for these specific fields, because the assumption is made they are always present and just popped out if not.

Better would be to have an action "read_declarations", which would then also be useable inside organization_show.
But this works for now.