add correct facet filters to organization detail page.
New filters (facets) were added and adjusted for the package search page, but these filters are also shown on the organization detail page. (in edit mode).

Make sure these filters are the same by:
adding organization_facets() call, which sets the facets for this page (set it the same as the dataset search facets)
Use get_facet_label_function automatically in facet_list implementation, so we don't need to explicitely pass it everytime we want to show the facets.
update bulk_process.html so it also adds the class "filters" (for some reason it is not part of the original code, so it doesn't follow the same structure as the other filter lists...)