Resource: show metadata modified date correctly
This should show the property metadata_modified. The rest is kept as a fallback, because it is unclear why it had to be this way in the past.
This might have been a data inconsistency in the past. Unclear if it can be removed, because historic commit messages don't give any information.

The text has also been changed to Metadata date as requested.