logo upload: show * for required
This is quite a hacky fix, because there is a lot of brittle code that made the <span> containing the label disappear.

Comment out the initialization code that sets the `field-logo-upload`. This already contains the expected text anyway.

Remove some custom logic that set the text. Unclear what the purpose was here.
It says it adjusts the label, but label is already correct...