fix normal user creating dataset
Because for creation, old value nap_checked is None, this would not be allowed as the user "tried to" set it to False.

Fix this by just keeping the old value for non-sysadmins: that way no special logic is needed to check what exactly the user send. The field is just ignored (not mandatory).

Also fix the code for checking _old_value: extras is an attribute, not a property.